requirements.txt
beautifulsoup4
requests

tqdm
markdownify

python-frontmatter
